app/requirements.txt
Flask==1.1.1
mysql-connector==2.2.9
pandas==1.1.3
names==0.3.0
flask-login==0.5.0
flask-wtf==0.14.3
pyopenssl==20.0.1
pymongo==3.11.4